anonymizer/requirements.txt
asttokens==2.4.1
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
fake-email==2.0.0
Faker==27.0.0
idna==3.10
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
numpy==1.24.4
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
pyarrow==17.0.0
